+++ src/de/mod.rs
@@ -1,5 +1,5 @@
 mod deserializer;
 mod packet_reader;
-pub mod packets;
+pub mod received_packet;
 pub use deserializer::Error;
 pub(crate) use packet_reader::PacketReader;
